major clean up
cleaned up a lot of junk and moved a couple pieces of logic around to
areas where they make more sense.

dropped the syslog handler because it's a nightmare to work with - i
might add it back at a later date if there's a need for it.

updated tests as needed.